requirements.txt
# espnet==202308
espnet_tts_frontend
faster-whisper==0.9.0
h5py
gradio==3.44.4
jaconv==0.3.4
kaldiio==2.18.0
librosa==0.10.1
# pyopenjtalk==0.3.2
pyopenjtalk-prebuilt==0.3.0
pyworld==0.3.4
sentencepiece==0.1.97
tensorboard==2.14
typeguard==2.13.3
torch_complex